projects/team-8/pyproject.toml
[project]
name = "pet-insurance-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "pandas>=2.0",
    "chromadb>=1.4.1",
    "fastapi>=0.128.5",
    "langchain>=0.2",
    "langchain-chroma>=0.1",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.2",
    "langchain-core>=0.2",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=0.2",
    "langchain-upstage>=0.7.6",
    "langgraph>=1.0.8",
    "langsmith>=0.6.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "rich>=14.3.2",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
    "kiwipiepy>=0.22.2",
]

[dependency-groups]
dev = [
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "httpx>=0.28.1",
    "ipython>=9.10.0",
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: real external dependency integration tests",
]
